project/requirements.txt
Django ==  4.2.2
django-crispy-forms== 2.0
djangorestframework ==3.14.0
Pillow ==9.5.0
psycopg2-binary == 2.9.6
python-dotenv  ==  1.0.0
pytz ==  2023.3
sqlparse  ==  0.4.4
typing_extensions  == 4.6.3
crispy-bootstrap5 == 0.7
gunicorn==20.1.0
